Remove MD5, SHA-1, & SHA-512 hashes & signatures
MD5 and SHA-1 are secure against preimage attacks. Even so, they probably
shouldn't be used any longer.

SHA-512 is just a 512-bit version of the SHA-2 standard, which we alraedy have
with SHA-256, which is still secure. MD5, SHA-1, and SHA-2 all use the
Merkle–Damgård construction, which could be a single point of failure if
weaknesses are found in that construction.

So removing MD5, SHA-1, and SHA-512 allows me to prepare for using hashes with
different constructions, avaiding that theoretical vulnerability.